refactor(explorer): move explorer routes into its own route group
