+++ packages/explorer/src/app/(explorer)/Providers.tsx
@@ -6,7 +6,7 @@ import { ReactNode } from "react";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { createConfig, http } from "@wagmi/core";
 import { localhost } from "@wagmi/core/chains";
-import { AppStoreProvider } from "../store";
+import { AppStoreProvider } from "../../store";
 
 const queryClient = new QueryClient();
