+++ packages/explorer/src/app/(explorer)/layout.tsx
