+++ packages/explorer/src/app/(explorer)/not-found.tsx
@@ -2,8 +2,8 @@
 
 import { ExternalLink } from "lucide-react";
 import Link from "next/link";
-import { Button } from "../components/ui/Button";
-import { useWorldUrl } from "../hooks/useWorldUrl";
+import { Button } from "../../components/ui/Button";
+import { useWorldUrl } from "../../hooks/useWorldUrl";
 
 export default function NotFound() {
   const getUrl = useWorldUrl();
